+++ tests/utils/test_date.py
@@ -1,6 +1,10 @@
 from datetime import datetime, timedelta
 
-from cg.utils.date import get_date_days_ago, get_timedelta_from_date
+from cg.utils.date import (
+    convert_string_to_datetime_object,
+    get_date_days_ago,
+    get_timedelta_from_date,
+)
 
 
 def test_get_date_days_ago(timestamp_now: datetime):
@@ -31,3 +35,34 @@ def test_get_timedelta_from_date(timestamp_yesterday: datetime):
 
     # Then the age in days should be yesterday
     assert age.days == 1
+
+
+def test_convert_string_to_datetime_object_valid():
+    # GIVEN a list of valid datetime strings with different formats
+    valid_datetime_strings = [
+        ("10/15/2024 12:45:30", "%m/%d/%Y %H:%M:%S"),  # Month/Day/Year, U.S. format
+    ]
+
+    # WHEN the function is called with valid strings
+    for date_str, expected_format in valid_datetime_strings:
+        # THEN it should return the correct datetime object
+        expected_datetime = datetime.strptime(date_str, expected_format)
+        assert convert_string_to_datetime_object(date_str) == expected_datetime
+
+
+def test_convert_string_to_datetime_object_invalid():
+    # GIVEN a list of invalid datetime strings
+    invalid_datetime_strings = [
+        "2024-15-10 12:45:30",  # Invalid day format
+        "15-10-2024 12:45",  # Missing seconds
+        "Invalid string",  # Completely invalid
+    ]
+
+    # WHEN the function is called with invalid strings
+    for date_str in invalid_datetime_strings:
+        # THEN it should raise a ValueError
+        try:
+            convert_string_to_datetime_object(date_str)
+            assert False, f"Expected ValueError for {date_str} but it didn't raise"
+        except ValueError as e:
+            assert str(e) == f"Could not convert '{date_str}' to a datetime object."
